Correct AppletFindById ABI

+++ os3k/os3k.h
@@ -227,7 +227,7 @@ char TranslateKeyToChar(KeyMod_e key);
 
 // These functions return an index of 0 (system applet) if not found.
 uint8_t AppletFindByName(char* name, uint8_t start_index);
-uint8_t AppletFindById(uint16_t id);
+uint32_t AppletFindById(uint16_t id);
 int AppletGetName(uint8_t index, char* name);
 int AppletSendMessage(uint8_t index, Message_e message, uint32_t param, uint32_t* status);
